EDGFQM-6 Integrate the RTR and system user changes
This commit integrates the recent changes in edge-common-spring that add
support for RTR and system users. There are no significant changes
needed in this module for these, but there were some breaking changes in
that library. To make it work, this commit adds a couple new properties
(including one dummy values that isn't actually used), it adds a new
depedendency (hibernate-validator), and fixes some tests.