EDGEFQM-16 Bump Spring Boot version
This commit bumps our Spring Boot version from 3.1.2 to 3.2.3, which is
the latest stable release. This also bumps our dependency versions for
folio-spring-base (7.1.2 -> 8.1.0), edge-common-spring (2.3.0 -> 2.3.2),
and edge-common (4.4.3 -> 4.5.2) to their latest releases, to bring them
all in alignment with the newer Spring Boot version

This also required adding a new dummy property
(folio.system-user.password) that is required for the module to start
up. This property isn't used.